backend/requirements.txt
fastapi
uvicorn
python-decouple
sqlalchemy
pymysql
sqlalchemy_utils